test: split copy tests
